refactor: extract PromptNSTextView to separate file
Extracts AppKit-specific text view code from PromptTextEditor.swift for better layer separation and code organization.

**Changes:**
- Created PromptNSTextView.swift with:
  - PromptNSTextView class (custom NSTextView with callback handlers)
  - calculatePromptContentHeight function (height calculation with line clamping)
- Updated PromptTextEditor.swift to remove duplicate code (75 lines removed)

**Benefits:**
- Clear separation: SwiftUI bridge vs AppKit implementation
- Improved modularity: AppKit code isolated in Support/ directory
- Maintains all existing behavior and tests (10/10 tests pass)
